requirements.txt
amqp==2.6.1
asgiref==3.3.1
billiard==3.6.3.0
boto3==1.16.15
botocore==1.19.15
celery==4.3.0
certifi==2020.11.8
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dicttoxml==1.7.4
Django==3.1.3
django-cors-headers==3.5.0
django-redis==4.12.1
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-api-key==2.0.0
drf-yasg==1.20.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
kombu==4.6.11
lxml==4.6.1
MarkupSafe==1.1.1
numpy==1.19.4
openapi-codec==1.3.2
packaging==20.4
psycopg2==2.7.5
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
redis==3.2.0
requests==2.24.0
requests-cache==0.5.2
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
sentry-sdk==0.16.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.11
vine==1.3.0
watchtower==1.0.0
xmljson==0.2.1
xmltodict==0.12.0
zipp==3.4.0
